feat(dm): jvan_battle_dmを整数(順位)からNumeric(5,1)(スコア)に変更
chunk[4:8]/10.0がスコア値（例:80.7）であることを確認し、
chunk[8:10]の順位から変更。DBカラム型も合わせてalter。

+++ backend/src/api/import_router.py
@@ -367,7 +367,7 @@ class JvanDmRecord(BaseModel):
     jravan_race_id: str  # 16文字 (例: "2026042503010501")
     horse_number: int
     jvan_time_dm: float | None = None
-    jvan_battle_dm: int | None = None
+    jvan_battle_dm: float | None = None
 
 
 class JvanDmRequest(BaseModel):
